Central agent

+++ src/ipc/ipc.c
@@ -1,21 +1,28 @@
 #include "ipc.h"
 #include "defines/ipc_values.h"
+#include "utils/ipUtils.h"
 #include "utils/logger.h"
 #include "utils/memory.h"
 #include "utils/oidc_error.h"
+#include "utils/printer.h"
+#include "utils/stringUtils.h"
 
+#include <arpa/inet.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <sys/ioctl.h>
 #include <sys/select.h>
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <unistd.h>
 
-oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer) {
-  con->server = secAlloc(sizeof(struct sockaddr_un));
-  con->sock   = secAlloc(sizeof(int));
+oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer,
+                                       int tcp) {
+  con->server     = secAlloc(sizeof(struct sockaddr_un));
+  con->tcp_server = secAlloc(sizeof(struct sockaddr_in));
+  con->sock       = secAlloc(sizeof(int));
   if (isServer) {  // msgsock is not needed for a client; furthermore if
                    // the client calls ipc_close it would close stdin
     con->msgsock = secAlloc(sizeof(int));
@@ -26,23 +33,25 @@ oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer) {
     return oidc_errno;
   }
 
-  *(con->sock) = socket(AF_UNIX, SOCK_STREAM, 0);
+  *(con->sock) = socket(tcp ? AF_INET : AF_UNIX, SOCK_STREAM, 0);
   if (*(con->sock) < 0) {
     logger(ERROR, "opening stream socket: %m");
     oidc_errno = OIDC_ECRSOCK;
     return oidc_errno;
   }
-  con->server->sun_family = AF_UNIX;
+  con->server->sun_family     = AF_UNIX;
+  con->tcp_server->sin_family = AF_INET;
   return OIDC_SUCCESS;
 }
-oidc_error_t initClientConnection(struct connection* con) {
-  return initConnectionWithoutPath(con, 0);
+
+oidc_error_t initClientConnection(struct connection* con, int tcp) {
+  return initConnectionWithoutPath(con, 0, tcp);
 }
 
 oidc_error_t initConnectionWithPath(struct connection* con,
                                     const char*        socket_path) {
   logger(DEBUG, "initializing ipc with path %s\n", socket_path);
-  if (initConnectionWithoutPath(con, 0) != OIDC_SUCCESS) {
+  if (initConnectionWithoutPath(con, 0, 0) != OIDC_SUCCESS) {
     return oidc_errno;
   }
   strcpy(con->server->sun_path, socket_path);
@@ -50,16 +59,13 @@ oidc_error_t initConnectionWithPath(struct connection* con,
 }
 
 /**
- * @brief initializes a client unix domain socket
+ * @brief initializes a client unix domain or tcp socket
  * @param con, a pointer to the connection struct. The relevant fields will be
  * initialized.
  * @param env_var_name, the socket_path environment variable name
  */
 oidc_error_t ipc_client_init(struct connection* con, const char* env_var_name) {
   logger(DEBUG, "initializing client ipc");
-  if (initClientConnection(con) != OIDC_SUCCESS) {
-    return oidc_errno;
-  }
   const char* path = getenv(env_var_name);
   if (path == NULL) {
     printError("Could not get the socket path from env var '%s'. Have you "
@@ -69,21 +75,49 @@ oidc_error_t ipc_client_init(struct connection* con, const char* env_var_name) {
            env_var_name);
     oidc_errno = OIDC_EENVVAR;
     return OIDC_EENVVAR;
+  }
+
+  int tcp = isValidIPOrHostnameOptionalPort(path);
+
+  if (initClientConnection(con, tcp) != OIDC_SUCCESS) {
+    return oidc_errno;
+  }
+
+  char* tmp_path = oidc_strcopy(path);
+  // split port and ip, check if ip part is ip and port part is number
+  char*          ip       = strtok(tmp_path, ":");
+  char*          port_str = strtok(NULL, ":");
+  unsigned short port     = port_str == NULL ? 0 : strToUShort(port_str);
+  if (tcp) {
+    logger(DEBUG, "Using TCP socket");
+    con->tcp_server->sin_port = htons(port ?: 42424);
+    con->tcp_server->sin_addr.s_addr =
+        inet_addr(isValidIP(ip) ? ip : hostnameToIP(ip));
   } else {
+    logger(DEBUG, "Using UNIX domain socket");
     strcpy(con->server->sun_path, path);
   }
+  secFree(tmp_path);
   return OIDC_SUCCESS;
 }
 
 /**
- * @brief connects to a UNIX Domain socket
+ * @brief connects to a UNIX Domain or TCP socket
  * @param con, the connection struct
  * @return the socket or @c OIDC_ECONSOCK on failure
  */
 int ipc_connect(struct connection con) {
-  logger(DEBUG, "connecting ipc %s\n", con.server->sun_path);
-  if (connect(*(con.sock), (struct sockaddr*)con.server,
-              sizeof(struct sockaddr_un)) < 0) {
+  struct sockaddr* server      = (struct sockaddr*)con.server;
+  size_t           server_size = sizeof(struct sockaddr_un);
+  if (con.server->sun_path[0] == '\0') {
+    server      = (struct sockaddr*)con.tcp_server;
+    server_size = sizeof(struct sockaddr_in);
+    logger(DEBUG, "connecting tcp ipc %lu:%hu\n",
+           con.tcp_server->sin_addr.s_addr, con.tcp_server->sin_port);
+  } else {
+    logger(DEBUG, "connecting ipc '%s'\n", con.server->sun_path);
+  }
+  if (connect(*(con.sock), server, server_size) < 0) {
     close(*(con.sock));
     logger(ERROR, "connecting stream socket: %m");
     oidc_errno = OIDC_ECONSOCK;
@@ -246,6 +280,8 @@ oidc_error_t ipc_closeConnection(struct connection* con) {
   }
   secFree(con->server);
   con->server = NULL;
+  secFree(con->tcp_server);
+  con->tcp_server = NULL;
   secFree(con->sock);
   con->sock = NULL;
   secFree(con->msgsock);
@@ -259,8 +295,10 @@ oidc_error_t ipc_closeConnection(struct connection* con) {
  * @return @c OIDC_SUCCESS on success
  */
 oidc_error_t ipc_closeAndUnlinkConnection(struct connection* con) {
-  logger(DEBUG, "Unlinking %s", con->server->sun_path);
-  unlink(con->server->sun_path);
+  if (con->server->sun_path) {
+    logger(DEBUG, "Unlinking %s", con->server->sun_path);
+    unlink(con->server->sun_path);
+  }
   ipc_closeConnection(con);
   return OIDC_SUCCESS;
 }
